requirements.txt
flask>=3.0
pillow>=10.0
anthropic>=0.40
gunicorn>=22.0
